+++ R/am.R
@@ -34,8 +34,8 @@ AM <- R6Class(
             if(nrow(self$etl) == 0L){
                 return(exclude.cols(eval(basic_stats), .escape = !pretty))
             }
-            eval(basic_stats)[eval(lkp_etl_logs), `:=`(meta = i.meta, last_event_time = i.timestamp, event = i.event, in_nrow = i.in_nrow, unq_nrow = i.unq_nrow, load_nrow = i.load_nrow)
-                              ][order(-last_event_time)
+            eval(basic_stats)[eval(lkp_etl_logs), `:=`(meta = i.meta, event_timestamp = i.timestamp, event = i.event, in_nrow = i.in_nrow, unq_nrow = i.unq_nrow, load_nrow = i.load_nrow)
+                              ][order(-event_timestamp)
                                 ][, exclude.cols(.SD, .escape = !pretty)
                                   ]
         },
@@ -130,13 +130,8 @@ AM <- R6Class(
             invisible(self)
         },
         # ETL
-        load = function(mapping, data, meta = NA_integer_, .args, use.im=TRUE){
+        load = function(mapping, data, meta = NA_integer_, use.im=TRUE){
             if(!isTRUE(private$instance_run)) stop("Run DW instance by am$run()")
-            if(!missing(.args)){
-                data <- .args[["data"]]
-                meta <- .args[["meta"]]
-                mapping <- .args[["mapping"]]
-            } # easier to call `load` programmatically
             stopifnot(
                 is.list(mapping),
                 length(mapping)>0L,
@@ -151,94 +146,79 @@ AM <- R6Class(
                stop(paste0("Invalid entity params provided, it should already be catched. A debug on `valid_entity_params`"))
             } # src columns "" / NULL exists in data to load # apply over elements in the mapping and then over names of each attribute definition
             model_all_attr_codes_for_anchors <- setNames(self$read(names(mapping))$childs, names(mapping))
-            model_attrs_lkp <- quote(self$read(unique(unlist(model_all_attr_codes_for_anchors)))[, .(code, hist, knot),, .(anchor, mne)])
-            mapping_attrs_dt <- rbindlist(lapply(names(mapping), function(nm) data.table(anchor = nm, mne = names(mapping[[nm]])[names(mapping[[nm]]) != ""]))) # exclude anchor NK
-            setkeyv(mapping_attrs_dt, c("anchor","mne"))
-            # TODO move first pass loop checks here? if knot provided, if hist provided
-            mapping_attrs_dt[ eval(model_attrs_lkp), `:=`(code = i.code, hist = i.hist, knot = i.knot)]
+            # key by hist and knot only for batch attributes match to model, by defined hist, knot, mne, anchor
+            model_attrs_lkp <- quote(self$read(unique(unlist(model_all_attr_codes_for_anchors)))[, .(code),, .(class, anchor, mne, hist, knot)])
+            # transform mapping to data.table
+            mapping_attrs_dt <- tryCatch(
+                rbindlist(lapply(names(mapping),
+                                 function(Aname) data.table(
+                                     class = "attribute",
+                                     anchor = Aname,
+                                     rbindlist(lapply(names(mapping[[Aname]])[names(mapping[[Aname]]) != ""], # exclude anchor NK
+                                                      function(aname) data.table(
+                                                          mne = aname,
+                                                          src_col = sapply(aname, function(anm) mapping[[Aname]][[anm]][1L]),
+                                                          hist = sapply(aname, function(anm) !is.na(as.character(as.list(mapping[[Aname]][[anm]])[["hist"]])[1L])),
+                                                          knot = sapply(aname, function(anm) as.character(as.list(mapping[[Aname]][[anm]])[["knot"]])[1L]),
+                                                          hist_col = sapply(aname, function(anm) as.character(as.list(mapping[[Aname]][[anm]])[["hist"]])[1L])
+                                                      )
+                                     ))
+                                 )
+                )),
+                warning = function(w) stop("Invalid mapping definition, see examples how to define mapping"),
+                error = function(e) stop("Invalid mapping definition, see examples how to define mapping")
+            )
+            setkeyv(mapping_attrs_dt, c("class","anchor","mne","hist","knot"))
+            # first pass loop checks on composite key join
+            mapping_attrs_dt[ eval(model_attrs_lkp), `:=`(code = i.code)]
             if(any(is.na(mapping_attrs_dt$code))){
-                stop(paste0("Some of the provided attributes do not exists in the model: ", paste(mapping_attrs_dt[is.na(code), paste(anchor, mne, sep="_")], collapse=", ")))
-            } # all provided attributes in the mapping exists in model for those anchors
+                stop(paste0("Some of the provided attributes have incorrect definition versus model: ", paste(mapping_attrs_dt[is.na(code), paste(anchor, mne, sep="_")], collapse=", ")))
+            } # all provided attributes in the mapping exists in model for those anchors, with expected hist and knot
             if(use.im){
                 data <- im$use(data, mne = mapping_attrs_dt[, unique(na.omit(c(anchor, knot)))], nk = lapply(mapping, `[[`, 1L), in.place = FALSE)
             } # auto Identity Management: anchors and knots get ID in incoming data and in am$IM() but not yet in obj$data
             # prepare sequence of processing
             load_seq <- setkeyv(rbindlist(
                 list(
-                    mapping_attrs_dt[, .(class = "anchor", mne = anchor, code = anchor, hist = FALSE, knot = NA_character_), .(anchor)],
-                    mapping_attrs_dt[!is.na(knot), .(anchor = NA_character_, class = "knot", mne = knot, code = knot, hist = FALSE, knot = NA_character_), .(byknot = knot)][, .SD, .SDcols=-"byknot"],
-                    mapping_attrs_dt[, .(anchor, class = "attribute", mne, code, hist, knot)]
-                )), c("anchor","class"))
+                    mapping_attrs_dt[, .(class = "anchor", mne = anchor, code = anchor, hist = FALSE, knot = NA_character_, src_col = paste_(anchor,"ID"), hist_col = NA_character_), .(anchor)],
+                    mapping_attrs_dt[!is.na(knot), .(class = "knot", anchor = NA_character_, mne = knot, code = knot, hist = FALSE, knot = NA_character_, src_col = NA_character_, hist_col = NA_character_), .(byknot = knot)][, .SD, .SDcols=-"byknot"],
+                    mapping_attrs_dt[, .(anchor, class = "attribute", mne, code, hist, knot, src_col = src_col, hist_col = hist_col)]
+                )), c("class","anchor"))
             set2keyv(load_seq, "code")
             # first pass loop, only check if mapping matches, fill defaults, etc.
-            for(load_code in load_seq$code){
-                browser()
-                # load_seq[code==load_code, !is.na(knot)]
-                iter <- self$read(code)[, .(mne,name,class,hist = sapply(obj, function(obj) isTRUE(obj$hist)),knot = sapply(obj, function(obj) as.character(obj$knot)[1L]),anchor = sapply(obj, function(obj) as.character(obj$anchor)[1L]))]
-                if(!is.na(iter$knot)){
-                    if(!iter$knot %chin% top_codes[class=="knot", code]) stop("Cannot load knotted attribute/tie without also loading knot for it, provide knot mapping.")
-                } # check if knotted
-                if(isTRUE(iter$hist)){
-                    if(iter$class=="attribute"){
-                        if(!"hist" %chin% names(mapping[[iter$anchor]][[iter$mne]])){
-                            stop(paste0("Cannot load historized ",iter$class," wihtout mapping of column on which historize ",iter$class,". Provide to the mapping 'hist' field for that ",iter$class," and give name of source dataset columns."))
-                        }
-                    }
-                    else if(iter$class=="tie"){
-                        if(!"hist" %chin% names(mapping[[iter$mne]])){
-                            stop(paste0("Cannot load historized ",iter$class," wihtout mapping of column on which historize ",iter$class,". Provide to the mapping 'hist' field for that ",iter$class," and give name of source dataset columns."))
-                        }
-                    }
-                } # check if historized must containt field "hist"
-                if(iter$class=="anchor"){
-                    if(length(mapping[[iter$mne]]) == 0L){
-                        if(!paste(iter$mne,"ID",sep="_") %chin% names(data)) stop(paste0("If anchor ID field was not provided it has to be in naming convention '",paste(iter$mne,"ID",sep="_"),"' (as IM class produces, see ?IM), otherwise provide anchor ID source column as unnamed character 'AC' into anchor mapping."))
-                        mapping[[iter$mne]][[paste_(iter$mne,"ID")]] <- paste_(iter$mne,"ID")
-                    } # use automapping for default naming convention column, [mne]_ID
-                    if(!all(sapply(names(mapping[[iter$mne]]), function(child_mne) is.null(names(mapping[[iter$mne]][[child_mne]])) || all(names(mapping[[iter$mne]][[child_mne]]) %chin% c("","anchors","anchor","roles","identifier","knot","hist"))))){
-                        stop(paste0('Element ',code,' in your mapping definition contains some childs elements which names do not match to allowed ',paste(c("","anchors","anchor","roles","identifier","knot","hist"),collapse=", ")))
-                    } # attributes nested in anchor character names validation
-                } # automapping anchor by name convention, check column names
-                if(iter$class=="knot"){
-                    if(length(mapping[[iter$mne]]) == 1L){
-                        if(!paste(iter$mne,"ID",sep="_") %chin% names(data)) stop(paste0("If knot ID field was not provided it has to be in naming convention '",paste(iter$mne,"ID",sep="_"),"' (as IM class produces, see ?IM), otherwise provide knot ID source in the data set by adding `c('GEN', id = 'GEN_ID')` into knot mapping."))
-                        mapping[[iter$mne]] <- c(mapping[[iter$mne]], id = paste(iter$mne,"ID",sep="_"))
-                    } # use automapping for default naming convention column, [mne]_ID
-                } # automapping for knots ID by name convention
-            }
-            browser()
-            # second pass loop, subset columns and provide to AMobj classes for loading
-            for(code in top_code){
-                # check if historized
-                if(isTRUE(self$read(mne)$obj[[1L]]$hist)){
-                    #if(length(mapping[[mne]]) < 2L || !"hist" %chin% names(mapping[[mne]])){
-                    #    #stop("Cannot load historized attribute wihtout mapping of columns on which historize attribute/tie. Provide historize source in the data set by adding `c('AC_NAM', hist = 'HireName')` into attribute/tie mapping.")
-                    #    # auto historize
-                    #    # mapping[[mne]][["hist"]] <- ".am"
-                    #    autohist <- TRUE
-                    #}
-                }
-
-                src_cols <- mapping[[mne]]
-                src_cols <- c(src_cols[is.null(names(src_cols))], src_cols[names(src_cols) %chin% c("id")], src_cols[names(src_cols) %chin% c("")], src_cols[names(src_cols) %chin% c("hist")]) # reorder cols
-                # check if historized
-                #if(isTRUE(self$read(mne)$obj[[1L]]$hist)){
-                #    src_cols <- c(src_cols,mapping[[mne]][["hist"]])
-                #}
-                tgt_cols <- self$read(mne)$obj[[1L]]$cols # already have hist if required
-                #if(length(self$read(mne)$obj[[1L]]$knot)){
-                #}
-                #self$ID[[mne]][data[, unique(.SD), .SDcols=self$NK[[mne]]]][is.na(eval(as.name(paste(mne,"ID",sep="_"))))]
-                #if(self$read(mne)$class=="attribute"){
-                #
-                #}
-                #if(length(src_cols)+1!=length(tgt_cols)) browser()
-                # subset cols, rename and load further
-                self$data[mne, obj][[1L]]$load(data = data[, src_cols, with=FALSE
-                                                           ][, c(.SD, if(isTRUE(autohist)) list(`.am.autohist` = rep(Sys.time(),.N)))
-                                                             ][, setnames(.SD, c(src_cols, if(isTRUE(autohist)) ".am.autohist"), tgt_cols[-length(tgt_cols)])],
-                                               meta = meta)
-            }
+            # ?
+            # loading knots
+            lapply(load_seq["knot", code, nomatch=0L], function(knot_code){
+                src_cols <- load_seq[c("attribute","tie"), nomatch=0L][knot==knot_code, src_cols]
+                src_cols <- c(paste_(knot_code,"ID"), src_cols)
+                cols <- self$read(knot_code)$obj[[1L]]$cols
+                cols <- cols[-length(cols)] # exclude metadata col
+                self$data[knot_code, obj][[1L]]$load(
+                    data = melt(data[, src_cols, with=FALSE], id.vars = src_cols[1L], measure.vars = src_cols[-1L], value.name = self$read(knot_code)$name)[, unique(.SD), .SDcols=-"variable"], # support for multi child knots, will load from all src_cols into one
+                    meta = meta
+                )
+            })
+            # loading anchors
+            lapply(load_seq["anchor", code, nomatch=0L], function(anchor_code){
+                src_cols <- load_seq[code==anchor_code, src_col]
+                cols <- self$read(anchor_code)$obj[[1L]]$cols
+                cols <- cols[-length(cols)] # exclude metadata col
+                self$data[anchor_code, obj][[1L]]$load(
+                    data = data[, src_cols, with=FALSE][, unique(.SD)],
+                    meta = meta
+                )
+                # loading child attributes
+                lapply(load_seq[CJ("attribute", anchor_code), code, nomatch=0L], function(attr_code){
+                    src_cols <- load_seq[code==attr_code, c(if(is.na(knot)) src_col else paste_(knot,"ID"), if(hist) hist_col else character())]
+                    src_cols <- c(paste_(anchor_code,"ID"), src_cols)
+                    cols <- self$read(attr_code)$obj[[1L]]$cols
+                    cols <- cols[-length(cols)] # exclude metadata col
+                    self$data[attr_code, obj][[1L]]$load(
+                        data = data[, src_cols, with=FALSE][, setnames(.SD, src_cols, cols)],
+                        meta = meta
+                    )
+                })
+            })
             invisible(self)
         },
         query = function(){
@@ -286,7 +266,7 @@ AM <- R6Class(
     ),
     active = list(
         log = function() rbindlist(private$log_list),
-        etl = function() rbindlist(lapply(self$data$obj, function(x) x$log)),
+        etl = function() setkeyv(rbindlist(lapply(self$data$obj, function(x) x$log)), c("meta","code")),
         add = function(){
             list("anchor" = function(...) self$create(class = "anchor", ...),
                  "A" = function(...) self$create(class = "anchor", ...),