+++ tests/testthat/test-loading.R
@@ -1,0 +1,16 @@
+context("AM loading method")
+
+test_that("Invalid mapping tests", {
+
+    # lack of hist
+    # lack of knot
+    # misspelled col name
+    # misspelled attr
+    # misspelled anchor
+    # col name invalid data type
+    # hist invalid data type
+    # knot invalid data type
+    # anchor invalid data type (non-list)
+    # check against defined naming convention for AM instance
+
+})